requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2